Randomize shop selection
The Generate Shop button creates a random assortment of items in the shop from a list
of all available items.  It currently repeats items.  This will need to be fixed.

+++ GameSquire/AddItemViewController.swift
@@ -23,7 +23,7 @@ class AddItemViewController: UIViewController {
     
     @IBAction func addBTN(_ sender: UIButton) {
         if Int(priceTF.text!) != nil && Int(quantityTF.text!) != nil{
-            AppDelegate.myModel.addItem(name: itemTF.text!, price: Int(priceTF.text!)!, description: descriptionTF.text!, quantity: Int(quantityTF.text!)!)
+            AppDelegate.myModel.addItem(name: itemTF.text!, price: Double(priceTF.text!)!, description: descriptionTF.text!, quantity: Int(quantityTF.text!)!)
         }
     }
